models and units
added code to ppmap to write out the model fit to the data and the residual map - at the moment only for the individual tiles.

added code to premap to read in a units parameter from premap to decide whether the units need to be changed.
Also changed readimage_wcs to read the BUNITS keyword from the fits headers to check for consistency, or use that if the premap parameter is not present.

also tinkered with the divider subroutine to make the behaviour a bit more intuitive, and changed ppmosaic a bit to make it work right when there is just one tile.