models and units
added code to ppmap to write out the model fit to the data and the residual map - at the moment only for the individual tiles.

added code to premap to read in a units parameter from premap to decide whether the units need to be changed.
Also changed readimage_wcs to read the BUNITS keyword from the fits headers to check for consistency, or use that if the prempa parameter is not present.